requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1